Reintroduce python3 in the Dockerfile
This PR reintroduces the changes made in previous PR's for moving the
Dockerfile and other files to python3. The breaking changes that were
discovered also seem to happen with previous images with python2. More
investigation is needed.

There is no backwards compatibility for python versions here, in case
such is needed a more sophisticated solution is needed.

References #80